pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "continuous-patterns"
version = "0.1.0"
description = "Research-oriented continuous-media pattern experiments (JAX)."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11,<3.14"
authors = [{ name = "continuous-patterns contributors" }]
dependencies = []

[project.optional-dependencies]
cpu = ["jax>=0.5"]
cuda = ["jax[cuda12]>=0.5"]
agate = [
    "jax[cuda12]>=0.5",
    "numpy>=2",
    "matplotlib>=3.8",
    "tqdm>=4",
    "h5py>=3",
    "scipy>=1.11",
    "pyyaml>=6",
    "imageio[ffmpeg]>=2",
]

[dependency-groups]
dev = [
    "pre-commit>=4",
    "pytest>=8",
    "ruff>=0.9",
    "scipy>=1.11",
    "tqdm>=4",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "slow: optional long-running GPU integration tests",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.format]
quote-style = "double"

[tool.uv]
default-groups = ["dev"]
